An attempt to migrate from neodark theme to something more supported
in a recent tree-sitter update, markdown colors became broken.
So I decided to migrate to something that is configurable and just apply
colors from neodark theme. The migration is not that smooth as I'd like it
to be, so I had to verride some things. The result is not perfect,
I don't like the style of visual selection (new version preserves fg,
just adds bg), but let's try in practice.